requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.5.0
magic-filter==1.0.12
multidict==6.1.0
packaging==24.2
propcache==0.2.0
pycron==3.1.1
pydantic==2.9.2
pydantic_core==2.23.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
sniffio==1.3.1
taskiq==0.11.9
taskiq-dependencies==1.5.4
taskiq-nats==0.5.1
taskiq-redis==1.0.2
typing_extensions==4.12.2
yarl==1.18.0
zipp==3.21.0
